requirements.txt
requests == 2.32.3
pyTelegramBotAPI == 4.24.0
python-dotenv == 1.0.1